requirements.txt
langchain==0.3.25
langchain-groq==0.3.2
langchain-community==0.3.24
langchain-huggingface==0.1.2
faiss-cpu>=1.9.0
sentence-transformers==3.4.1
pydantic==2.11.3
python-dotenv==1.1.0
streamlit==1.44.1
tiktoken==0.9.0
unstructured==0.17.2
markdown==3.7
streamlit
